chore: Update Dockerfile for containerized deployment

+++ src/utils/redis.ts
@@ -1,20 +1,31 @@
 import {createClient} from 'redis';
-import {REDIS_URL} from "../server_env.ts";
-
-const client = await createClient({url: REDIS_URL}).on('error', (err) => {
-    console.log(err);
-}).connect().catch((err) => {
-    console.log(err);
-});
+import {REDIS_URL,ENABLE_REDIS} from "../server_env.ts";
+let client = null;
+if(ENABLE_REDIS){
+    client = await createClient({url: REDIS_URL}).on('error', (err) => {
+        console.log(err);
+    }).connect().catch((err) => {
+        console.log(err);
+    });
+}
 
 export async function get(key: string): Promise<string> {
-    return await client.get(key) as string;
+    if(ENABLE_REDIS){
+        return await client.get(key) as string;
+    }
+    return null;
 }
 
 export async function set(key: string, value: string, ttl?: number): Promise<void> {
-    await client.set(key, value, ttl);
+    if(ENABLE_REDIS){
+        await client.set(key, value, ttl);
+    }
+    return;
 }
 
 export async function del(key: string): Promise<void> {
-    await client.del(key);
+    if(ENABLE_REDIS){
+        await client.del(key);
+    }
+    return;
 }
